Extract separate module for navigation styles
This cleans up `app/routes` by removing style configuration, and
resolves a type inference error when type checking (due to the
differences in structure between the iOS and Android configurations).
